requirements.txt
numpy==1.12.0
setuptools==34.3.1
tensorflow_gpu==1.0.1
Pillow==4.0.0
six==1.10.0
